requirements.txt

langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf
sentence-transformers
faiss-cpu
streamlit
python-dotenv
transformers
torch
accelerate